Changing type

+++ MinimedKit/PumpModel.swift
@@ -51,7 +51,7 @@ public enum PumpModel: String {
         return (generation >= 23) ? 40 : 10
     }
 
-    var reservoirCapacity: Double {
+    var reservoirCapacity: Int {
         switch size {
         case 5:
             return 176
